+++ domains/kriegspiel.h
@@ -396,7 +396,7 @@ namespace GTLib2 {
             // Destructor
             ~KriegspielState() override = default;
 
-            int countAvailableActionsFor(Player player) const override;
+            unsigned long countAvailableActionsFor(Player player) const override;
 
             // GetActions returns possible actions for a player in the state.
             vector<shared_ptr<Action>> getAvailableActionsFor(Player player) const override;
