+++ domains/phantomTTT.h
@@ -95,7 +95,7 @@ class PhantomTTTState : public State {
   // Destructor
   ~PhantomTTTState() override = default;
 
-  int countAvailableActionsFor(Player player) const override;
+  unsigned long countAvailableActionsFor(Player player) const override;
 
   // GetActions returns possible actions for a player in the state.
   vector<shared_ptr<Action>> getAvailableActionsFor(Player player) const override;
